[IMP] implementation of all the config buttons
Ended the implementation of all the needed buttons in the CONFIG page.

Still has a refresh issue, will create a github issue asking for fix.

all the buttons are working but the main problem is still waiting fix. more info in #91

had a double call of all the buttons. no idea why. fixed.

Sleep will not be permitted in the forseeable future.

Luca.